requirements.txt
google-genai
streamlit
pypdf
firebase-admin

